Moved ring code to optional module
