+++ src/main/java/magicbees/client/gui/ContainerEffectRing.java
@@ -4,7 +4,7 @@ import net.minecraft.entity.player.InventoryPlayer;
 
 import cpw.mods.fml.common.registry.GameRegistry;
 import forestry.core.gui.ContainerItemInventory;
-import magicbees.itemInventories.InventoryBeeRing;
+import magicbees.main.utils.compat.baubles.InventoryBeeRing;
 
 public class ContainerEffectRing extends ContainerItemInventory<InventoryBeeRing> {
