+++ src/main/java/magicbees/client/gui/GUIHandler.java
@@ -4,7 +4,7 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
 
 import cpw.mods.fml.common.network.IGuiHandler;
-import magicbees.itemInventories.InventoryBeeRing;
+import magicbees.main.utils.compat.baubles.InventoryBeeRing;
 import magicbees.tileentity.TileEntityEffectJar;
 import magicbees.tileentity.TileEntityMagicApiary;
