+++ src/main/java/magicbees/main/utils/compat/baubles/RingHousing.java
@@ -1,4 +1,4 @@
-package magicbees.tileentity;
+package magicbees.main.utils.compat.baubles;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
@@ -20,7 +20,6 @@ import forestry.api.core.EnumTemperature;
 import forestry.api.core.IErrorLogic;
 import forestry.api.core.IErrorState;
 import magicbees.bees.BeeManager;
-import magicbees.itemInventories.InventoryBeeRing;
 
 // Class that simulates a BeeHouse using both player data and info from InventoryBeeRing
 public class RingHousing implements IBeeHousing {
